backend/requirements.txt
fastapi
uvicorn[standard]
motor
httpx[http2]
aiohttp
pydantic
python-multipart
python-jose[cryptography]
requests
weasyprint
jinja2
python-dotenv
python-dateutil
sse-starlette
pytest
pytest-asyncio
pytest-cov
black
isort
flake8
pdfminer.six
python-docx
PyMuPDF
click
